requirements.txt
# Il framework web
Flask==3.0.0

# L'adattatore per far parlare Python con PostgreSQL
# Usiamo -binary perché è più facile da installare in Docker (non richiede compilatori C)
psycopg2-binary==2.9.9

# (Opzionale ma consigliato) Per gestire il database in modo più semplice tramite oggetti (ORM)
Flask-SQLAlchemy==3.1.1

# (Opzionale) Per gestire variabili d'ambiente in modo ordinato
python-dotenv==1.0.0